FriendshipsTest FriendshipsFeedRequest - moved to friendships package, refactored, javadoc FriendshipsFeedResponse - moved to friendships package, refactored
