+++ lib/src/main/java/io/github/taz03/jia/responses/friendships/FriendshipsFeedResponse.java
@@ -1,18 +1,18 @@
-package io.github.taz03.jia.responses.users;
+package io.github.taz03.jia.responses.friendships;
+
+import io.github.taz03.jia.responses.InstagramResponse;
+import java.util.List;
+import io.github.taz03.jia.responses.models.user.User;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.github.taz03.jia.responses.InstagramResponse;
-import io.github.taz03.jia.responses.models.user.User;
-
-import java.util.List;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class FriendshipsFeedResponse extends InstagramResponse {
     @JsonProperty("users")
     private List<User> users;
 
-    public List getUsers() {
+    public List<User> getUsers() {
         return users;
     }
 }
